+++ src/generator/write-files.mjs
@@ -1,0 +1,1 @@
+export * from '../../scripts/writer.mjs'